+++ package.json
@@ -22,6 +22,7 @@
     "sass": "^1.54.9",
     "typescript": "^4.8.3",
     "vite": "^3.1.0",
+    "vite-plugin-glsl": "^0.3.0",
     "vue-tsc": "^0.40.13"
   }
 }
