+++ src/game/_resources/programs/unlit/unlit.frag
@@ -1,0 +1,11 @@
+#version 300 es
+
+precision mediump float;
+
+uniform vec3 u_color;
+
+out vec4 FragColor;
+
+void main() {
+  FragColor = vec4(u_color, 1.0);
+}
