+++ src/game/extensions.d.ts
@@ -1,0 +1,9 @@
+declare module '*.vert' {
+  const value: string;
+  export default value;
+}
+
+declare module '*.frag' {
+  const value: string;
+  export default value;
+}
