+++ vite.config.ts
@@ -1,10 +1,11 @@
-import {defineConfig} from 'vite';
-import {visualizer} from 'rollup-plugin-visualizer';
 import vue from '@vitejs/plugin-vue';
+import {visualizer} from 'rollup-plugin-visualizer';
+import {defineConfig} from 'vite';
+import glsl from 'vite-plugin-glsl';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [vue(), glsl()],
   server: {
     hmr: false,
   },
